runc: split the lib's trait from lib to async and sync
